+++ mvrx/src/main/kotlin/com/airbnb/mvrx/MavericksTestOverrides.java
@@ -3,7 +3,6 @@ package com.airbnb.mvrx;
 
 import androidx.annotation.RestrictTo;
 
-@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 @InternalMavericksApi
 public class MavericksTestOverrides {
     /**